Refactor CI pipeline
