+++ .circleci/config.yml
@@ -23,7 +23,8 @@ jobs:
             make lint
             black --check --exclude tools/file_packager.py .
             mypy --ignore-missing-imports pyodide_build/ src/ packages/micropip/micropip/ packages/*/test*
-  build:
+
+  build-core:
     <<: *defaults
     steps:
       - checkout
@@ -37,7 +38,61 @@ jobs:
           command: sudo apt install -y libtinfo5
 
       - run:
-          name: build
+          name: build emsdk
+          no_output_timeout: 1200
+          command: |
+            ccache -z
+            make -C emsdk
+            ccache -s
+
+      - run:
+          name: build cpython
+          no_output_timeout: 1200
+          command: |
+            ccache -z
+            make -C cpython
+            ccache -s
+
+      - run:
+          name: build pyodide core
+          no_output_timeout: 1200
+          command: |
+            ccache -z
+            # the following packages are currently used in the main pyodide test suite
+            PYODIDE_PACKAGES="micropip,pyparsing,pytz,packaging,kiwisolver" make
+            ccache -s
+
+      - run:
+          name: check-size
+          command: ls -lh build/
+
+      - save_cache:
+          paths:
+            - ./emsdk/emsdk
+            - ~/.ccache
+          key: v1-emsdk-{{ checksum "emsdk/Makefile" }}-v20201203-{{ .BuildNum }}
+
+      - persist_to_workspace:
+          root: .
+          paths:
+            - .
+
+  build-packages:
+    <<: *defaults
+    steps:
+      - attach_workspace:
+          at: .
+
+      - restore_cache:
+          keys:
+          - v1-emsdk-{{ checksum "emsdk/Makefile" }}-v20201203-
+
+      - run:
+          name: dependencies
+          command: sudo apt install -y libtinfo5
+
+      - run:
+          name: build packages
           no_output_timeout: 1200
           command: |
             ccache -z
@@ -63,63 +118,45 @@ jobs:
       - store_artifacts:
           path: /home/circleci/repo/build/
 
-  build-test-minimal:
+  test-core-firefox:
     <<: *defaults
     steps:
-      - checkout
-
-      # this cache is generated by the main build job, we never store it here
-      - restore_cache:
-          keys:
-          - v1-emsdk-{{ checksum "emsdk/Makefile" }}-v20201203-
-
-      - run:
-          name: dependencies
-          command: sudo apt install -y libtinfo5
-
-      - run:
-          name: build
-          no_output_timeout: 1200
-          command: |
-            ccache -z
-            PYODIDE_PACKAGES='micropip' make
-            ccache -s
-
-      - run:
-          name: check-size
-          command: ls -lh build/
-
-      - run:
-          name: test
-          command: |
-            export PYODIDE_PACKAGES='micropip'
-            # only check common tests, all other are checked in the main test jobs
-            pytest src/tests/test_common.py -v -k firefox
-
-      - store_artifacts:
-          path: /home/circleci/repo/build/
-
-  test-firefox:
-    <<: *defaults
-    steps:
-      - checkout
       - attach_workspace:
           at: .
       - run:
           name: test
           command: |
-            pytest src packages/*/test* packages/micropip/micropip/ pyodide_build -v -k firefox
+            pytest src packages/micropip/ -v -k firefox
 
-  test-chrome:
+  test-core-chrome:
     <<: *defaults
     steps:
-      - checkout
       - attach_workspace:
           at: .
       - run:
           name: test
           command: |
-            pytest src packages/*/test* packages/micropip/micropip/ pyodide_build -v -k chrome
+            pytest src packages/micropip/ -v -k chrome
+
+  test-packages-firefox:
+    <<: *defaults
+    steps:
+      - attach_workspace:
+          at: .
+      - run:
+          name: test
+          command: |
+            pytest packages/*/test* -v -k firefox
+
+  test-packages-chrome:
+    <<: *defaults
+    steps:
+      - attach_workspace:
+          at: .
+      - run:
+          name: test
+          command: |
+            pytest packages/*/test* -v -k chrome
 
   test-python:
     <<: *defaults
@@ -204,21 +241,30 @@ workflows:
   build-and-deploy:
     jobs:
       - lint
-      - build:
+      - build-core:
           requires:
             - lint
           filters:
             tags:
               only: /.*/
-      - build-test-minimal:
+      - build-packages:
           requires:
-            - lint
-      - test-chrome:
+            - buld-core
+      - test-core-chrome:
           requires:
-            - build
-      - test-firefox:
+            - build-core
+      - test-core-firefox:
           requires:
-            - build
+            - build-core
+          filters:
+            tags:
+              only: /.*/
+      - test-packages-chrome:
+          requires:
+            - build-packages
+      - test-packages-firefox:
+          requires:
+            - build-packages
           filters:
             tags:
               only: /.*/
@@ -240,7 +286,8 @@ workflows:
               only: /^\d+\.\d+\.\d+$/
       - deploy-s3:
           requires:
-            - test-firefox
+            - test-core-firefox
+            - test-packages-firefox
             - test-python
           filters:
             branches:
